+++ CHANGELOG.md
@@ -1,5 +1,14 @@
+# v0.2.0
+##  15-02-2020
+
+1. [](#new)
+    * Option to pretty-print
+    * Option to only return name: version
+2. [](#improved)
+    * Code-quality
+
 # v0.1.0
-##  25/11/2017
+##  25-11-2017
 
 1. [](#new)
     * Initial beta-release